Aligned the create team enterprise form wit the billing upgrade form
