+++ frontend/src/pages/team/changeType.vue
@@ -1,7 +1,7 @@
 <template>
     <ff-page>
         <ff-loading v-if="loading" message="Updating Team..." />
-        <div v-else class="m-auto">
+        <div v-else>
             <form class="space-y-6">
                 <div>
                     <FormHeading>Change your team type</FormHeading>
@@ -76,7 +76,6 @@
 
 <script>
 import { ChevronLeftIcon } from '@heroicons/vue/outline'
-import { defineComponent } from 'vue'
 import { mapState } from 'vuex'
 
 import billingApi from '../../api/billing.js'
@@ -84,18 +83,21 @@ import instanceTypesApi from '../../api/instanceTypes.js'
 import teamApi from '../../api/team.js'
 import teamTypesApi from '../../api/teamTypes.js'
 import FormHeading from '../../components/FormHeading.vue'
+import { useHubspotHelper } from '../../composables/Hubspot.js'
 
 import Alerts from '../../services/alerts.js'
-import Dialog from '../../services/dialog.js'
 import Product from '../../services/product.js'
-
-import ObjectProperties from './Brokers/Hierarchy/components/schema/ObjectProperties.vue'
 
 // eslint-disable-next-line vue/one-component-per-file
 export default {
     name: 'ChangeTeamType',
     components: {
         FormHeading
+    },
+    setup () {
+        const { talkToSalesCalendarModal } = useHubspotHelper()
+
+        return { talkToSalesCalendarModal }
     },
     data () {
         return {
@@ -308,43 +310,7 @@ export default {
             }
         },
         sendContact: async function () {
-            const user = this.user
-            Dialog.show({
-                header: 'Talk to sales to Upgrade to Enterprise',
-                kind: 'primary',
-                boxClass: 'ff-dialog-box--wide',
-                confirmLabel: 'Close',
-                canBeCanceled: false,
-                is: {
-                    // eslint-disable-next-line vue/one-component-per-file
-                    component: defineComponent({
-                        components: { ObjectProperties },
-                        computed: {
-                            url () {
-                                const url = new URL('flowfuse/book-a-demo-call', 'https://meetings-eu1.hubspot.com')
-                                url.searchParams.set('embed', true)
-                                url.searchParams.set('firstName', user.name.split(' ')[0])
-                                url.searchParams.set('lastName', user.name.split(' ')[1] ?? '')
-                                url.searchParams.set('email', user.email)
-
-                                return url.toString()
-                            }
-                        },
-                        mounted () {
-                            // Create and inject the HubSpot script tag
-                            const script = document.createElement('script')
-                            script.type = 'text/javascript'
-                            script.src = 'https://static.hsappstatic.net/MeetingsEmbed/ex/MeetingsEmbedCode.js'
-                            script.async = true
-                            document.body.appendChild(script)
-                        },
-                        template: `
-                            <p>Unlock advanced features, dedicated support, and enterprise scale capabilities. Schedule a call to discuss your needs.</p>
-                            <div class="meetings-iframe-container" :data-src="url"></div>
-                          `
-                    })
-                }
-            })
+            this.talkToSalesCalendarModal(this.user)
         }
     }
 }
